Moving display resolution to separate module
