+++ brewtils/decorators.py
@@ -2,14 +2,10 @@
 
 import functools
 import inspect
-import json
-import os
 import sys
-from io import open
 from types import MethodType
 from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Type, Union
 
-import requests
 import six
 
 try:
@@ -18,6 +14,7 @@ except ImportError:
     from lark.common import ParseError
 
 from brewtils.choices import parse
+from brewtils.display import resolve_display_modifiers
 from brewtils.errors import PluginParamError, _deprecate
 from brewtils.models import Command, Parameter, Choices
 
@@ -443,7 +440,7 @@ def _initialize_command(method):
     cmd.name = _method_name(method)
     cmd.description = cmd.description or _method_docstring(method)
 
-    resolved_mod = _resolve_display_modifiers(
+    resolved_mod = resolve_display_modifiers(
         method, cmd.name, schema=cmd.schema, form=cmd.form, template=cmd.template
     )
     cmd.schema = resolved_mod["schema"]
@@ -493,77 +490,6 @@ def _method_docstring(method):
         docstring = method.__doc__
 
     return docstring.split("\n")[0] if docstring else None
-
-
-def _resolve_display_modifiers(
-    wrapped,  # type: MethodType
-    command_name,  # type: str
-    schema=None,  # type: Union[dict, str]
-    form=None,  # type: Union[dict, list, str]
-    template=None,  # type: str
-):
-    # type: (...) -> dict
-    """Parse display modifier parameter attributes
-
-    Returns:
-        Dictionary that fully describes a display specification
-    """
-
-    def _load_from_url(url):
-        response = requests.get(url)
-        if response.headers.get("content-type", "").lower() == "application/json":
-            return json.loads(response.text)
-        return response.text
-
-    def _load_from_path(path):
-        current_dir = os.path.dirname(inspect.getfile(wrapped))
-        file_path = os.path.abspath(os.path.join(current_dir, path))
-
-        with open(file_path, "r") as definition_file:
-            return definition_file.read()
-
-    resolved = {}
-
-    for key, value in {"schema": schema, "form": form, "template": template}.items():
-
-        if isinstance(value, six.string_types):
-            try:
-                if value.startswith("http"):
-                    resolved[key] = _load_from_url(value)
-
-                elif value.startswith("/") or value.startswith("."):
-                    loaded_value = _load_from_path(value)
-                    resolved[key] = (
-                        loaded_value if key == "template" else json.loads(loaded_value)
-                    )
-
-                elif key == "template":
-                    resolved[key] = value
-
-                else:
-                    raise PluginParamError(
-                        "%s specified for command '%s' was not a "
-                        "definition, file path, or URL" % (key, command_name)
-                    )
-            except Exception as ex:
-                raise PluginParamError(
-                    "Error reading %s definition from '%s' for command "
-                    "'%s': %s" % (key, value, command_name, ex)
-                )
-
-        elif value is None or (key in ["schema", "form"] and isinstance(value, dict)):
-            resolved[key] = value
-
-        elif key == "form" and isinstance(value, list):
-            resolved[key] = {"type": "fieldset", "items": value}
-
-        else:
-            raise PluginParamError(
-                "%s specified for command '%s' was not a definition, "
-                "file path, or URL" % (key, command_name)
-            )
-
-    return resolved
 
 
 def _sig_info(arg):
